Add migration to replicate existing default workspaces
